CLOG-32 : Configuration bug when using logback along with chronicle-logg...

+++ logger/src/main/java/net/openhft/chronicle/logger/VanillaLogAppenderConfig.java
@@ -123,8 +123,7 @@ public class VanillaLogAppenderConfig extends ChronicleLogAppenderConfig {
         return this.builder.dataBlockSize();
     }
 
-    //TODO: long vs int
-    public void setDataBlockSize(int dataBlockSize) {
+    public void setDataBlockSize(long dataBlockSize) {
         this.builder.dataBlockSize(dataBlockSize);
     }
